lessons/lesson.38/web-app/pyproject.toml
[tool.poetry]
name = "web-app"
version = "0.1.0"
description = ""
authors = ["Suren Khorenyan <surenkhorenyan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "<2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.0.1"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
